# openapi_spec.yaml
openapi: 3.0.3
info:
  title: roll_with_the_punches_ecommerce
  version: 1.0.0
  description: >-
    Roll with the Punches E-Commerce is an online retailer for all of the best
    boxing gear
  contact:
    name: Reed McDaniel
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /users/{id}:
    summary: Specific user by ID
    get:
      summary: Get a user by ID
      description: Get a user by ID
      operationId: get_user
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - users
    patch:
      summary: Update a user by ID
      description: Update a user by ID
      operationId: patch_user
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - users
    delete:
      summary: Delete user by ID
      description: Delete user by ID
      operationId: delete_user
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - users
...
  /products/{id}:
    summary: Specific product by ID
    get:
      summary: Get a specific product by ID
      description: Get a specific product by ID
      operationId: get_product
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - products
    patch:
      summary: Update a product by ID
      description: Update a product by ID
      operationId: patch_product
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - products
    delete:
      summary: Delete product by ID
      description: Delete product by ID
      operationId: delete_product
      parameters:
        - $ref: "#/components/parameters/idParam"
      tags:
        - products
...
components:
  parameters:
    idParam:
      name: id
      description: resource ID
      in: path
      required: true
      schema:
        type: string
...